migrate tour types
